Factor non-Agent related functions out
This takes newly created non-actore related functions and puts them into a separate file.
