+++ src/acl.ts
@@ -469,3 +469,48 @@ export const internal_accessModeIriStrings = {
 } as const;
 /** @internal */
 type AccessModeIriString = typeof internal_accessModeIriStrings[keyof typeof internal_accessModeIriStrings];
+
+/** @internal
+ * This function finds, among a set of ACL rules, the ones granting access to a given entity (the target)
+ * and identifying it with a specific predicate (`acl:agent` or `acl:agentGroup`).
+ * @param aclRules The set of rules to filter
+ * @param targetIri The IRI of the target
+ * @param targetType The predicate linking the rule to the target
+ */
+export function internal_getAclRulesForIri(
+  aclRules: unstable_AclRule[],
+  targetIri: IriString,
+  targetType: typeof acl.agent | typeof acl.agentGroup
+): unstable_AclRule[] {
+  return aclRules.filter((rule) =>
+    getIriAll(rule, targetType).includes(targetIri)
+  );
+}
+
+/** @internal
+ * This function transforms a given set of rules into a map associating the IRIs
+ * of the entities to which permissions are granted by these rules, and the permissions
+ * granted to them. Additionnally, it filters these entities based on the predicate
+ * that refers to them in the rule.
+ */
+export function internal_getAccessByIri(
+  aclRules: unstable_AclRule[],
+  targetType: typeof acl.agent | typeof acl.agentGroup
+): Record<IriString, unstable_Access> {
+  const targetIriAccess: Record<IriString, unstable_Access> = {};
+
+  aclRules.forEach((rule) => {
+    const ruleTargetIri = getIriAll(rule, targetType);
+    const access = internal_getAccess(rule);
+
+    // A rule might apply to multiple agents. If multiple rules apply to the same agent, the Access
+    // Modes granted by those rules should be combined:
+    ruleTargetIri.forEach((targetIri) => {
+      targetIriAccess[targetIri] =
+        typeof targetIriAccess[targetIri] === "undefined"
+          ? access
+          : internal_combineAccessModes([targetIriAccess[targetIri], access]);
+    });
+  });
+  return targetIriAccess;
+}
